Added cache to PersistenceManager
Added a static cache property to the PersistenceManager to
make calls to the file system less frequent.  getCycle
method now checks the cache for a cycle before going to the
file system.  If the cycle is not in the cache, getCycle
gets it from the file system then adds it to the cache.

saveCycle method automatically adds the cycle it is saving
to the cache.

Resolves #9